requirements.txt
numpy
scikit-learn
logzero
easydict
tqdm
yacs
termcolor
ray==1.2.0
h5py
matplotlib
Pillow
torch==1.6.0
torchvision==0.7.0
pandas
nltk==3.5
librosa
torch-audiomentations

git+https://github.com/wbaek/theconf.git@v0.1.5
git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git
git+https://github.com/makcedward/nlpaug.git